requirements.txt
alembic==1.13.1
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
filelock==3.14.0
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fsspec==2024.5.0
future==1.0.0
greenlet==3.0.3
huggingface-hub==0.23.2
idna==3.7
intel-openmp==2021.4.0
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
Mako==1.3.5
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
psycopg2==2.9.9
PyAudio==0.2.14
pygame==2.5.2
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.2
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
sentence-transformers==3.0.0
serial==0.0.97
SpeechRecognition==3.10.4
SQLAlchemy==2.0.30
sympy==1.12.1
tbb==2021.12.0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.0
tqdm==4.66.4
transformers==4.41.2
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.3
WTForms==3.1.2
